add scheduler

+++ nile-application/src/main/kotlin/tech/volkov/nile/dto/MetricDto.kt
@@ -3,6 +3,6 @@ package tech.volkov.nile.dto
 data class MetricDto(
     val name: String,
     val description: String,
-    val scrapeInterval: Int,
+    val scrapeInterval: Long,
     val sampleValue: Double
 )
